+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 - Updated supported runtimes for functions. Default python runtime is now 3.13. (#8534)
 - Changed artifact registry cleanup policy error to warn for CI/CD workloads #8513
+- Enhance firebase init apphosting to support local source deploys. (#8479)
+- Add GCP API client functions to support App Hosting deploy from source feature. (#8545)